requirements.txt
pytesseract
Pillow
transformers
pymupdf>=1.27.1